Add a system of unity
add system of unity with notes of 1/16 beat. With the ui it is now impossible to make a modification of less than 1/16 of beat
1 beat = 1 second, I will see for making a system of tempo
BUG REPORT : when lot  of edits made to the window, when we play the file and it scroll automatically, the sound is bad quality. When it stops scrolling the sound become clear again. If we save the music and reload it and play again, the sound is completly clear. Planning to make a purge of all variable at each play but may take more time to play then.